Split files for easier testing
This will allow us to export anything we'd like from the main.js file, while reserving the index file to export ONLY public methods.
