+++ package.json
@@ -39,10 +39,12 @@
   },
   "devDependencies": {
     "@types/inquirer": "^6.5.0",
+    "@types/jest": "^24.0.20",
     "@types/node": "^12.7.2",
     "@types/semver": "^6.0.1",
     "eslint": "^6.0.1",
     "inquirer": "^7.0.0",
+    "jest": "^24.9.0",
     "rollup": "^1.16.6",
     "runscript": "^1.4.0",
     "semver": "^6.3.0",
